+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- New: Added Track and Attachments columns for the logbook table. Shows if a flight record has a track attached and amount of the attachments. Hidden by default
+
 ## [3.8.0] - 06.09.2025
 
 - New: Added a support to change the naming for the standard fields on the flight record page and all the tables (logbook and stats)