+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Fix: Don't show the brackets if the person doesn't have a role
 - Update: Change Map selectors `No Route Lines`/`No Tracks` to `Route Lines`/`Tracks`
 - New: Added Track and Attachments columns for the logbook table. Shows if a flight record has a track attached and amount of the attachments. Hidden by default